+++ CHANGELOG.md
@@ -6,3 +6,4 @@ All notable changes to this project will be documented in this file.
 
 - Homepage Hero component
 - Bees Tiers component
+- Hard Truth component